requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
instaloader==4.14
load-dotenv==0.1.0
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
python-telegram-bot==21.8
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3
validators==0.34.0
yarl==1.18.3
yt-dlp==2024.12.3
